Changes based on Mike's feedback.

+++ src/hal/sensor/sensor.h
@@ -6,12 +6,12 @@
 #include "../../hal/hal.h"
 
 // Initializes the pressure and airflow sensors
-int16_t sensorHalInit(void);
+int sensorHalInit(void);
 
 // Gets the airflow sensor value
-int16_t airflowSensorHalGetValue(int16_t *value);
+int airflowSensorHalGetValue(int16_t *value);
 
 // Gets the pressor sensor value (measured in units of 0.1mmH2O
-int16_t pressureSensorHalGetValue(int16_t *value);
+int pressureSensorHalGetValue(int16_t *value);
 
 #endif /* __SENSOR_HAL_H__ */
